DataBackend for SpatRaster

A mlr3::DataBackend for SpatRaster (package terra).

Read mode

Block mode reads complete rows of the raster file and subsets the requested cells. Faster than cell mode if we iterate the whole raster file.

Cell mode reads individual cells. Faster than block mode if only a few cells are sampled.

Block mode is activated if $data(rows) is called with a increasing integer sequence e.g. 200:300.

Super class

mlr3::DataBackend -> DataBackendSpatRaster

Public fields

compact_seq

logical(1)
If TRUE, row ids are a natural sequence from 1 to nrow(data) (determined internally). In this case, row lookup uses faster positional indices instead of equi joins.

Active bindings

rownames

(integer())
Returns vector of all distinct row identifiers, i.e. the contents of the primary key column.

colnames

(character())
Returns vector of all column names, including the primary key column.

nrow

(integer(1))
Number of rows (observations).

ncol

(integer(1))
Number of columns (variables), including the primary key column.

stack

(integer(1))
Returns SpatRaster.

Methods

Public methods

DataBackendSpatRaster$new()

DataBackendSpatRaster$data()

DataBackendSpatRaster$head()

DataBackendSpatRaster$distinct()

DataBackendSpatRaster$missings()

Inherited methods

mlr3::DataBackend$format()

mlr3::DataBackend$print()

Method new()

Creates a backend for a SpatRaster.

Usage

DataBackendSpatRaster$new(data)

Arguments

data

(SpatRaster)
A raster object.

Method data()

Returns a slice of the data. Calls raster::rowColFromCell() and raster::getValues() on the spatial object and converts it to a data.table::data.table().

The rows must be addressed as vector of primary key values, columns must be referred to via column names. Queries for rows with no matching row id and queries for columns with no matching column name are silently ignored. Rows are guaranteed to be returned in the same order as rows, columns may be returned in an arbitrary order. Duplicated row ids result in duplicated rows, duplicated column names lead to an exception.

Usage

DataBackendSpatRaster$data(rows, cols, data_format = "data.table")

Arguments

rows

integer()
Row indices. Row indices start with 1 in the upper left corner in the raster, increase from left to right and then from top to bottom. The last cell is in the bottom right corner and the row index equals the number of cells in the raster.

cols

character()
Column names.

data_format

(character(1))
Desired data format, e.g. "data.table" or "Matrix".

Method head()

Retrieve the first n rows.

Usage

DataBackendSpatRaster$head(n = 6L)

Arguments

n

(integer(1))
Number of rows.

Returns

data.table::data.table() of the first n rows.

Method distinct()

Returns a named list of vectors of distinct values for each column specified. If na_rm is TRUE, missing values are removed from the returned vectors of distinct values. Non-existing rows and columns are silently ignored.

Usage

DataBackendSpatRaster$distinct(rows, cols, na_rm = TRUE)

Arguments

rows

integer()
Row indices. Row indices start with 1 in the upper left corner in the raster, increase from left to right and then from top to bottom. The last cell is in the bottom right corner and the row index equals the number of cells in the raster.

cols

character()
Column names.

na_rm

logical(1)
Whether to remove NAs or not.

Returns

Named list() of distinct values.

Method missings()

Returns the number of missing values per column in the specified slice of data. Non-existing rows and columns are silently ignored.

Usage

DataBackendSpatRaster$missings(rows, cols)

Arguments

rows

integer()
Row indices. Row indices start with 1 in the upper left corner in the raster, increase from left to right and then from top to bottom. The last cell is in the bottom right corner and the row index equals the number of cells in the raster.

cols

character()
Column names.

Returns

Total of missing values per column (named numeric()).

Examples

if (mlr3misc::require_namespaces("terra", quietly = TRUE)) {
  stack = demo_stack_spatraster(size = 5, layers = 5)
  backend = DataBackendSpatRaster$new(stack)
}