Predict on spatial objects with mlr3 learners

This function allows to directly predict mlr3 learners on various spatial objects (see section "Supported Spatial Classes"). It returns an mlr3::Prediction object and (optionally) the same object that was used for the prediction.

predict_spatial_newdata(learner, object, filename = NULL, overwrite = FALSE)

Arguments

learner
mlr3::Learner
Any mlr3::Learner.

object
[SpatRaster, sf, RasterBrick]

filename
[character]
Filename of optional file to write prediction values into. For raster-like inputs this can be a .tif file. For sf objects, this could be a .gpgk or a .shp file.

overwrite
[logical]
Should a possibly existing file on disk (referring to argument filename) be overwritten?

Value

mlr3::Prediction

Details

A direct prediction on a subset of a mlr3::Task object is not possible for terra objects as terra objects contain external pointers which are not compatible with future-based parallelization. Due to this, the values from the terra object need to be extracted first into a data.table.

Parallelization

For predictions which take > 10 seconds, parallelization could help speeding things up. mlr3 supports parallel predictions since v0.12.0. This can be enabled by setting the $parallel_predict = TRUE flag in the learner and supplying a parallel future plan before executing the prediction, for example future::plan(multisession, workers = 2). See the examples for more information.

Spatial Classes support

Task and Prediction support for the following classes is planned:

sf

stars

raster

Examples

if (mlr3misc::require_namespaces(c("terra", "future"), quietly = TRUE)) {
  stack = demo_stack_spatraster(size = 5, layers = 5)
  backend = DataBackendSpatRaster$new(stack)
  task = as_task_classif(backend, target = "y", positive = "TRUE")
  # train
  learner = lrn("classif.featureless")
  learner$train(task, row_ids = sample(1:task$nrow, 500))
  predict_spatial_newdata(learner, stack)
}

#> <PredictionClassif> for 250000 observations:
#>     row_ids truth response
#>           1 FALSE    FALSE
#>           2 FALSE    FALSE
#>           3 FALSE    FALSE
#> ---                       
#>      249998  TRUE    FALSE
#>      249999  TRUE    FALSE
#>      250000  TRUE    FALSE