mlr3spatial: Support for Spatial Objects Within the 'mlr3' Ecosystem

Extends the 'mlr3' ML framework with methods for spatial objects. Data storage and prediction are supported for packages 'terra', 'raster' and 'stars'.

Learn mlr3

Book on mlr3: https://mlr3book.mlr-org.com

Use cases and examples gallery: https://mlr3gallery.mlr-org.com

Cheat Sheets: https://github.com/mlr-org/mlr3cheatsheets

mlr3 extensions

Preprocessing and machine learning pipelines: mlr3pipelines

Analysis of benchmark experiments: mlr3benchmark

More classification and regression tasks: mlr3data

Connector to OpenML: mlr3oml

Solid selection of good classification and regression learners: mlr3learners

Even more learners: https://github.com/mlr-org/mlr3extralearners

Tuning of hyperparameters: mlr3tuning

Hyperband tuner: mlr3hyperband

Visualizations for many mlr3 objects: mlr3viz

Survival analysis and probabilistic regression: mlr3proba

Cluster analysis: mlr3cluster

Feature selection filters: mlr3filters

Feature selection wrappers: mlr3fselect

Interface to real (out-of-memory) data bases: mlr3db

Performance measures as plain functions: mlr3measures

Suggested packages

Parallelization framework: future

Progress bars: progressr

Encapsulated evaluation: evaluate, callr (external process)

Package Options

"mlr3.debug": If set to TRUE, parallelization via future is disabled to simplify debugging and provide more concise tracebacks. Note that results computed with debug mode enabled use a different seeding mechanism and are not reproducible.

"mlr3.allow_utf8_names": If set to TRUE, checks on the feature names are relaxed, allowing non-ascii characters in column names. This is an experimental and temporal option to pave the way for text analysis, and will likely be removed in a future version of the package. analysis.

References

Becker M, Schratz P (2022). mlr3spatial: Support for Spatial Objects Within the 'mlr3' Ecosystem. https://mlr3spatial.mlr-org.com, https://github.com/mlr-org/mlr3spatial.

See also

Author

Maintainer: Marc Becker marcbecker@posteo.de (ORCID)

Authors:

Patrick Schratz patrick.schratz@gmail.com (ORCID)